Sanitized http_rest_proxy API (MLDB-1396)
