+++ sdk/sdk_include_seed.cc
@@ -7,6 +7,8 @@
     This file should include the list of headers that are part of the
     public C++ SDK.  These files, and header files that they depend on,
     will be installed as the API exposed by the SDK.
+
+    None of these may include boost headers, directly or transiently.
 */
 
 #include "mldb/core/dataset.h"
@@ -14,4 +16,15 @@
 #include "mldb/core/function.h"
 #include "mldb/core/procedure.h"
 #include "mldb/types/basic_value_descriptions.h"
-
+#include "mldb/vfs/filter_streams.h"
+#include "mldb/vfs/fs_utils.h"
+#include "mldb/http/http_rest_proxy.h"
+#include "mldb/rest/poly_entity.h"
+#include "mldb/sql/expression_value.h"
+#include "mldb/types/any_impl.h"
+#include "mldb/types/any.h"
+#include "mldb/types/date.h"
+#include "mldb/types/json_parsing.h"
+#include "mldb/types/json_printing.h"
+#include "mldb/types/structure_description.h"
+#include "mldb/types/value_description.h"
